Refactor linkmap language selection based on review feedback
- Extract common getClangId() logic into rex_linkmap_clang_aware_trait
- Extract context handling into rex_linkmap_context_aware_trait
- Remove code duplication in renderer classes and concrete implementations
- Add dropdown-menu-right support for proper alignment in narrow windows
- Improve rex_view::clangSwitchAsDropdown() with showEditLink parameter

Addresses review feedback from PR #6298:
- Eliminates method duplication across abstract classes
- Consolidates common context/clang logic into reusable traits
- Maintains clean separation of concerns
