+++ redaxo/src/addons/structure/lib/linkmap/linkmap.php
@@ -7,22 +7,11 @@
  */
 class rex_linkmap_category_tree extends rex_linkmap_tree_renderer
 {
+    use rex_linkmap_context_aware_trait;
+
     public function __construct(
         private rex_context $context,
     ) {}
-
-    public function getContext(): rex_context
-    {
-        return $this->context;
-    }
-
-    /**
-     * Get the language ID from context.
-     */
-    protected function getClangId(): int
-    {
-        return (int) $this->context->getParam('clang', rex_clang::getStartId());
-    }
 
     /**
      * @return string
@@ -59,22 +48,11 @@ class rex_linkmap_category_tree extends rex_linkmap_tree_renderer
  */
 class rex_linkmap_article_list extends rex_linkmap_article_list_renderer
 {
+    use rex_linkmap_context_aware_trait;
+
     public function __construct(
         private rex_context $context,
     ) {}
-
-    public function getContext(): rex_context
-    {
-        return $this->context;
-    }
-
-    /**
-     * Get the language ID from context.
-     */
-    protected function getClangId(): int
-    {
-        return (int) $this->context->getParam('clang', rex_clang::getStartId());
-    }
 
     /**
      * @return string
